+++ router/router_test.go
@@ -4,10 +4,11 @@ import (
 	"net/http"
 	"testing"
 
-	controllersMocks "github.com/ditrit/badaas/mocks/controllers"
-	middlewaresMocks "github.com/ditrit/badaas/mocks/router/middlewares"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
+
+	controllersMocks "github.com/ditrit/badaas/mocks/controllers"
+	middlewaresMocks "github.com/ditrit/badaas/mocks/router/middlewares"
 )
 
 func TestSetupRouter(t *testing.T) {
@@ -15,7 +16,7 @@ func TestSetupRouter(t *testing.T) {
 	middlewareLogger := middlewaresMocks.NewMiddlewareLogger(t)
 	authenticationMiddleware := middlewaresMocks.NewAuthenticationMiddleware(t)
 
-	basicController := controllersMocks.NewBasicAuthentificationController(t)
+	basicController := controllersMocks.NewBasicAuthenticationController(t)
 	informationController := controllersMocks.NewInformationController(t)
 	jsonController.On("Wrap", mock.Anything).Return(func(response http.ResponseWriter, request *http.Request) {})
 	router := SetupRouter(jsonController, middlewareLogger, authenticationMiddleware, basicController, informationController)